requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
PyPDF2==3.0.1
python-docx==0.8.11
nltk==3.8.1
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.25.2
requests==2.31.0
Pillow==10.1.0
gunicorn==21.2.0
psycopg2-binary==2.9.7
whitenoise==6.6.0
dj-database-url==2.1.0